Check if the LDAP server is available before connecting
Before connecting with ldap_connect and ldap_bind to the LDAP server,
the server will be pinged in order to ensure that it is available.

If this is not done, the connection will wait until it times out
(typically after 30 seconds). This way of implementing it is needed
because the PHP LDAP library does not provide a way to specify a
timeout.