Don't know what to put here, so just read the description please
Interesting changes are from lines 54 to 59. Other changes are just
cosmetic ones.

Before this change, FR3DLdapBundle instantiate a new instance of the
user class and hydrates it with the attributes of the Ldap.

With this change, the bundle tries to retrieve an existing user before
and fallback to instantiating a new user if it fails. This prevents, for
example, the FOSUserBundle to create a new user instead of updating the
existing one when using the profile edit feature.